refactor to library
