+++ src/pini.cpp
@@ -1,10 +1,11 @@
-#include "pini.hpp"
 #include <cstddef>
 #include <cstdint>
 #include <sstream>
 #include <string_view>
 #include <vector>
-#include "util.hpp"
+#include <pini/pini.hpp>
+#include <util.hpp>
+
 
 namespace pn {
 namespace {
